cleaned finance package
