+++ feature/finance/src/main/kotlin/org/mifospay/feature/finance/navigation/FinanceNavigation.kt
@@ -11,14 +11,17 @@ const val FINANCE_ROUTE = "finance_route"
 fun NavController.navigateToFinance(navOptions: NavOptions) = navigate(FINANCE_ROUTE, navOptions)
 
 fun NavGraphBuilder.financeScreen(
-    onAddBtn: () -> Unit
+    onAddBtn: () -> Unit,
+    onLevel1Clicked: () -> Unit,
+    onLevel2Clicked: () -> Unit,
+    onLevel3Clicked: () -> Unit
 ) {
     composable(route = FINANCE_ROUTE) {
         FinanceRoute(
             onAddBtn = onAddBtn,
-            onLevel1Clicked = {},
-            onLevel2Clicked = {},
-            onLevel3Clicked = {}
+            onLevel1Clicked = onLevel1Clicked,
+            onLevel2Clicked = onLevel2Clicked,
+            onLevel3Clicked = onLevel3Clicked,
         )
     }
 }
