--- mifospay/src/main/java/org/mifospay/home/navigation/FinanceNavigation.kt
@@ -1,27 +1,0 @@
-package org.mifospay.home.navigation
-
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavOptions
-import androidx.navigation.compose.composable
-import org.mifospay.home.screens.FinanceRoute
-
-const val FINANCE_ROUTE = "finance_route"
-
-fun NavController.navigateToFinance(navOptions: NavOptions) = navigate(FINANCE_ROUTE, navOptions)
-
-fun NavGraphBuilder.financeScreen(
-    onAddBtn: () -> Unit,
-    onLevel1Clicked: () -> Unit,
-    onLevel2Clicked: () -> Unit,
-    onLevel3Clicked: () -> Unit
-) {
-    composable(route = FINANCE_ROUTE) {
-        FinanceRoute(
-            onAddBtn = onAddBtn,
-            onLevel1Clicked = onLevel1Clicked,
-            onLevel2Clicked = onLevel2Clicked,
-            onLevel3Clicked = onLevel3Clicked
-        )
-    }
-}
