+++ mifospay/src/main/java/org/mifospay/navigation/MifosNavHost.kt
@@ -10,6 +10,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.mifospay.core.model.domain.Transaction
 import org.mifospay.common.Constants
+import org.mifospay.feature.finance.navigation.financeScreen
 import org.mifospay.feature.home.navigation.HOME_ROUTE
 import org.mifospay.feature.home.navigation.homeScreen
 import org.mifospay.feature.kyc.navigation.kycLevel1Screen
@@ -32,7 +33,6 @@ import org.mifospay.feature.savedcards.navigation.navigateToAddCard
 import org.mifospay.feature.settings.navigation.navigateToSettings
 import org.mifospay.feature.settings.navigation.settingsScreen
 import org.mifospay.feature.specific.transactions.SpecificTransactionsActivity
-import org.mifospay.home.navigation.financeScreen
 import org.mifospay.home.navigation.paymentsScreen
 import org.mifospay.payments.send.navigation.navigateToSendMoneyScreen
 import org.mifospay.payments.send.navigation.sendMoneyScreen
